Move block generator to own file in framework dir;
